requirements.txt
pandas==2.2.2
openpyxl==3.1.5
scikit-learn==1.4.2
pydantic==2.8.2
google-genai==1.0.0
fastapi==0.115.6
uvicorn[standard]==0.30.6
python-multipart==0.0.12
twilio==9.3.6
jinja2>=3.0.0
python-dotenv>=1.0.0